requirements.txt
Flask==3.0.3
matplotlib==3.8.3
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
pyngrok==7.1.6
pytorch_lightning==2.2.1
scikit_learn==1.4.1.post1
torch==2.2.1
torchmetrics==1.3.2
torchvision==0.17.1
openai==1.25.1
gunicorn==19.9.0
